support reading/writing of >4GB chunks (#91)
* use 64 bit numbers when dealing with TypedArrays

* old versions of GDAL have a problem with this test

* better sparse file

* test all read() cases

* do not replace the original Node.js exception

* include a >4G elements test

* clean up debug printfs

* the error message is Node.js version dependent

* replace the cryptic Node.js error message

* further improve the tests and fix data_type

* accept both type and data_type

* clarify this change

* relax codecov patch requirements

* reduce the memory requirement of the test

* actually test the data

+++ src/utils/typed_array.hpp
@@ -45,11 +45,11 @@ namespace node_gdal {
 
 namespace TypedArray {
 
-Local<Value> New(GDALDataType type, unsigned int length);
-Local<Value> New(GDALDataType type, void *data, unsigned int length);
+Local<Value> New(GDALDataType type, int64_t length);
+Local<Value> New(GDALDataType type, void *data, int64_t length);
 GDALDataType Identify(Local<Object> array);
-void *Validate(Local<Object> obj, GDALDataType type, int min_length);
-bool ValidateLength(int length, int min_length);
+void *Validate(Local<Object> obj, GDALDataType type, int64_t min_length);
+bool ValidateLength(size_t length, int64_t min_length);
 } // namespace TypedArray
 
 } // namespace node_gdal
